pytestframework/requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
appdirs==1.4.4
Appium-Python-Client==1.0.2
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
coverage==5.2
cryptography==2.9.2
distlib==0.3.0
execnet==1.7.1
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.5.1
geventhttpclient==1.4.2
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpath==0.82
locust==1.0.2
logger==1.4
MarkupSafe==1.1.1
more-itertools==8.3.0
msgpack==1.0.0
packaging==20.4
paramiko==2.7.1
pipenv==2020.6.2
pluggy==0.13.1
psutil==5.7.0
py==1.8.1
pycparser==2.20
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
pytest==5.4.2
pytest-forked==1.1.3
pytest-rerunfailures==9.0
pytest-xdist==1.32.0
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
sshtunnel==0.1.5
urllib3==1.25.9
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
